+++ src/matvar_cell.c
@@ -110,7 +110,7 @@ Mat_VarGetCells(matvar_t *matvar,int *start,int *stride,int *edge)
         cnt[1]++;
         idx[1] += stride[1];
         for ( j = 1; j < matvar->rank; j++ ) {
-            if ( cnt[j] == edge[j] ) {
+		  if ( (int)cnt[j] == edge[j] ) {
                 cnt[j] = 0;
                 idx[j] = start[j];
                 cnt[j+1]++;
